requirements.txt
Django==2.2.3
django-rest-framework==0.1.0
djangorestframework==3.9.4
gunicorn==19.9.0
pytz==2019.1
six==1.12.0
sqlparse==0.3.0
